% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grades.R
\name{adj_grades_moodle}
\alias{adj_grades_moodle}
\title{Adjust Grade}
\usage{
adj_grades_moodle(data_cleaned, new_max_grade = NULL, round_digits = 3)
}
\arguments{
\item{data_cleaned}{A cleaned data.frame}

\item{new_max_grade}{Length 1 numeric vector indicate new maximum grade, If \code{NULL} use old maximum grade}

\item{round_digits}{Length 1 numeric vector indicate digits to round grade, If \code{NULL} no rounding.}
}
\value{
A data.frame
}
\description{
Adjust Grade
}
