% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sub.R
\name{check_sub}
\alias{check_sub}
\title{Check Student Submission}
\usage{
check_sub(
  data,
  extract_id_from = c("Email address", "Institution", "Department", "ID number"),
  id_regex = ".*",
  sep_name = " ",
  state = c("Finished", "In progress"),
  encode = c(1, 0),
  choose_encode = c("max", "min", "all"),
  choose_time = c("first", "last", "all"),
  ...
)
}
\arguments{
\item{data}{A data.frame \strong{or} named list of data.frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Quiz report(s)} (i.e. either Grades or Responses report).}

\item{extract_id_from}{(Character) Choose 1 column to extract ID from}

\item{id_regex}{(Character) A regular expression used to extract ID from column (choose one) "Email address", "Institution", "Department", or "ID number" in the Moodle Quiz report. The default is "\verb{.*}" meaning all characters.
If your student email addresses has numeric IDs in them, try "\verb{[:digit:]+}" to extract digits from the email.
\strong{Note}: Regular expression syntax is the same as \href{https://github.com/rstudio/cheatsheets/blob/master/strings.pdf}{stringr}.}

\item{sep_name}{A character in the new "Name" column that separate original "First name" and "Surname".}

\item{state}{A character vector to match values in "State" column of the Moodle Quiz report}

\item{encode}{An encoding numeric vector corresponding to \code{state}.
For example: by default, in the "State" column, "Finished" values will be encoded as \code{1}, and
"In progress" will be encoded as \code{0}.}

\item{choose_encode}{A character to filter student's attempt by the \code{encode}ing.
\itemize{
\item \strong{"max"} (default): return rows that have maximum encoding of each students.
\item \strong{"min"}: return rows that have minimum encoding of each students.
\item \strong{"all"}: no filter applied, return all rows.
}}

\item{choose_time}{A character to filter student's attempt by started time (determined by "Started on" column in Moodle Quiz report).
This filter applies \strong{after} \code{choose_encode} has been applied to the data.
\itemize{
\item \strong{"first"} (default): return rows that represent first attempt of each students.
\item \strong{"last"}: return rows that represent last attempt of each students.
\item \strong{"all"}: no filter applied, return all rows after applying \code{choose_encode}
}}

\item{...}{argument \code{sep_col} of \code{check_sub.list()}, which indicates a character separation between names of list and "state" or "encode" columns.}
}
\value{
\strong{A data.frame}, its output content is determined by class of its first argument: \code{data}.
\itemize{
\item If the \code{data} is a data.frame; the output is an encoded, filtered, and cleaned data.frame of Moodle Quiz report.
\item If the \code{data} is a named list of data.frame; the output is the same as previously described, but all Moodle Quiz reports are \href{https://dplyr.tidyverse.org/reference/mutate-joins.html}{full-joined} together by column "Name" and "ID".
And, a new column "Total" is computed by the sum of all "...State" columns.
}
}
\description{
\code{\link[=check_sub]{check_sub()}} is a generic function that check student submission(s) from \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Quiz report(s)}
(i.e. Grades or Responses report). This function can encode & filter student's attempts by state (i.e., "Finished" or "In progress") and started time ("Started on" column).
This function also cleans column names of Moodle Quiz report for easier manipulation, extracts student ID from "Email address" column using regular expression,
and unites "First name" and "Surname" column into "Name".
}
\examples{
# Submission of Data Frame
check_sub(grades_ls$Quiz_1,
          id_regex = "[:digit:]+")

# Submission of List of DF
check_sub(grades_ls,
          id_regex = "[:digit:]+")


}
