% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_moodle.R
\name{clean_moodle}
\alias{clean_moodle}
\title{Clean Moodle Quiz report}
\usage{
clean_moodle(
  data,
  extract_id = TRUE,
  id_regex = ".*",
  sep_name = " ",
  dash_na = TRUE,
  force_numeric = TRUE
)
}
\arguments{
\item{data}{A data.frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Quiz report}.}

\item{extract_id}{(Logical) \code{TRUE}: An "ID" column will be created by extracting characters from "Email address" column
using regular expression as \code{id_regex}. If \code{FALSE}: "Email address" column will be renamed to "Email".}

\item{id_regex}{(Character) Regular expression used to extract ID from "Email address". The default is ".*" meaning all characters.
If your student email addresses has numeric IDs in them, try "\verb{[:digit:]+}" to extract digits from the email.
\strong{Note}: Regular expression syntax is the same as \href{https://github.com/rstudio/cheatsheets/blob/master/strings.pdf}{stringr}.}

\item{sep_name}{(Character) A character in the new "Name" column that separate original "First name" and "Surname".}

\item{dash_na}{(Logical) \code{TRUE}: format dash "-" in any columns to \code{NA}. If \code{FALSE}: leave dash "-" as is}

\item{force_numeric}{(Logical) \code{TRUE}: force "Grade" and "Q. x /x" columns to numeric type (if any of "Not yet graded", "Requires grading", or "-" is presented it will be \code{NA}).
If \code{FALSE}: leave "Grade" and "Q. x /x" columns as is.}
}
\value{
A data.frame with cleaner column names and better formatted data.
Although the column names are cleaned, some metadata is lost such as maximum score of quiz and questions.
Check out \code{\link[=get_quiz_attr]{get_quiz_attr()}} for get these metadata.
}
\description{
Clean Moodle Quiz report for easier manipulation in \code{R}. It extracts student ID from "Email address" column;
unites "First name" and "Surname" column; converts dash ("-"), which can be presented when student not given answer, to \code{NA};
and forces "Grade" and "Q. x /x" to numeric value (if any of "Not yet graded", "Requires grading", or "-" is presented it will be \code{NA}).
}
