% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer.R
\name{cloze}
\alias{cloze}
\title{Succinctly create a suitable cloze question}
\usage{
cloze(x, ...)
}
\arguments{
\item{x}{The correct answer}

\item{...}{Options passed on to other methods}
}
\description{
Succinctly create a suitable cloze question
}
