% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grades.R
\name{combine_grades}
\alias{combine_grades}
\title{Combine Grades}
\usage{
combine_grades(
  data,
  extract_id_from = c("Email address", "Institution", "Department"),
  id_regex = ".*",
  sep_name = " ",
  new_max_grade = NULL,
  round_digits = 3,
  choose_grade = c("max", "min", "mean", "all"),
  choose_time = c("first", "last", "all"),
  force_grade = F,
  sep_col = "_"
)
}
\arguments{
\item{data}{A data.frame \strong{or} named list of data.frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Grades report(s)} (not Responses report)}

\item{extract_id_from}{(Character) Choose 1 column to extract ID from}

\item{id_regex}{(Character) A regular expression used to extract ID from column "Email address" in the Moodle Quiz report. The default is "\verb{.*}" meaning all characters.
If your student email addresses has numeric IDs in them, try "\verb{[:digit:]+}" to extract digits from the email.
\strong{Note}: Regular expression syntax is the same as \href{https://github.com/rstudio/cheatsheets/blob/master/strings.pdf}{stringr}.}

\item{sep_name}{A character in the new "Name" column that separate original "First name" and "Surname".}

\item{new_max_grade}{(Numeric) A desired new maximum grade(s) to adjust. If \code{NULL} (default): the new maximum grade is/are equal to the old maximum grade from the Moodle Grades report (in "Grade/xx" columns).
\strong{Note}: The new maximum grade will be appended at the "Grade_*" column names.
\itemize{
\item If the \code{data} is a \strong{data.frame}; enter \code{new_max_grade} as length 1 numeric vector to indicate the new maximum grade of that quiz.
\item If the \code{data} is a \strong{named list of data.frame}; enter \code{new_max_grade} as numeric vector to indicate new maximum grades corresponding to each quizzes (the element of list of data.frame).
Length of \code{new_max_grade} must equal to length of list of data.frame.
}}

\item{round_digits}{Length 1 numeric vector to indicate digits to round grades. If \code{NULL}: no round digits.}

\item{choose_grade}{A character to filter student's attempt by score ("Grade/xx" column), useful when quiz has multiple attempts for each students.
\itemize{
\item \strong{"max"} (default): return rows that have maximum score of each students.
\item \strong{"min"}: return rows that have minimum score of each students.
\item \strong{"mean"}: compute and return mean score of each students.
\item \strong{"all"}: no filter applied, return all rows.
}}

\item{choose_time}{A character to filter student's attempt by started time (determined by "Started on" column).
This filter applies \strong{after} \code{choose_grade} has been applied to the data.
\itemize{
\item \strong{"first"} (default): return rows that represent first attempt of each students.
\item \strong{"last"}: return rows that represent last attempt of each students.
\item \strong{"all"}: no filter applied, return all rows after applying \code{choose_grade}
}}

\item{force_grade}{(Logical) If grade of any students was found to be "Not yet graded",
you will get an error message. The purpose of this is to warn you that you might have forgotten to grade some students.
If you want to bypass this behavior, set \code{force_grade = TRUE}.}

\item{sep_col}{(Character) If \code{data} is a named list of data.frame, \code{sep_col} indicate a character separation between names of list and "State" or "Grade" columns.}
}
\value{
\strong{A data.frame}, its output content is determined by class of its first argument: \code{data}.
\itemize{
\item If the \code{data} is a data.frame; the output is a cleaned data.frame of Moodle Grades report, whether it was adjusted or filtered depends on other arguments.
\item If the \code{data} is a named list of data.frame; the output is the same as previously described, but all Moodle Grades reports are \href{https://dplyr.tidyverse.org/reference/mutate-joins.html}{full-joined} together by column "Name" and "ID".
So that, "Grades_" columns from each data.frame are sit together in a single data.frame, and "Total_" column is added at the last column to indicate sum of all quizzes grades.
}
}
\description{
\code{\link[=combine_grades]{combine_grades()}} is a generic function that combine, adjust, and filter student grades from \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Grades report(s)} (not Responses report).
This function can combine "Grade/xx" columns from multiple Moodle Grades reports into a single data.frame.
Furthermore, you can adjust or weight maximum score of each quizzes and "Total" score will be computed.
If the report contains multiple attempts per student, you can also filter grades of each student by score ("Grade/xx" column) and start time ("Started on" column), e.g. the first best score of each students.
Like \code{\link[=check_sub]{check_sub()}}, this function also cleans column names for easy manipulation, extracts student ID from "Email address", and unites "First name" and "Surname" column into "Name".
}
\examples{
NULL
}
