% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_resp.R
\name{combine_resp}
\alias{combine_resp}
\title{Combine Student Responses}
\usage{
combine_resp(
  data,
  id_regex = ".*",
  sep_name = " ",
  state = c("Finished", "In progress"),
  encode = c(1, 0),
  choose_encode = c("max", "min", "all"),
  choose_time = c("first", "last", "all"),
  split_cloze = F,
  part_glue = "_part_",
  sep_col = "_"
)
}
\arguments{
\item{data}{A data.frame \strong{or} named list of data.frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Responses report(s)} (not Grades report)}

\item{id_regex}{(Character) A regular expression used to extract ID from column "Email address" in the Moodle Quiz report. The default is "\verb{.*}" meaning all characters.
If your student email addresses has numeric IDs in them, try "\verb{[:digit:]+}" to extract digits from the email.
\strong{Note}: Regular expression syntax is the same as \href{https://github.com/rstudio/cheatsheets/blob/master/strings.pdf}{stringr}.}

\item{sep_name}{A character in the new "Name" column that separate original "First name" and "Surname".}

\item{state}{A character vector to match values in "State" column of the Moodle Quiz report}

\item{encode}{An encoding numeric vector corresponding to \code{state}.
For example: by default, in the "State" column, "Finished" values will be encoded as \code{1}, and
"In progress" will be encoded as \code{0}.}

\item{choose_encode}{A character to filter student's attempt by the \code{encode}ing.
\itemize{
\item \strong{"max"} (default): return rows that have maximum encoding of each students.
\item \strong{"min"}: return rows that have minimum encoding of each students.
\item \strong{"all"}: no filter applied, return all rows.
}}

\item{choose_time}{A character to filter student's attempt by started time (determined by "Started on" column in Moodle Quiz report).
This filter applies \strong{after} \code{choose_encode} has been applied to the data.
\itemize{
\item \strong{"first"} (default): return rows that represent first attempt of each students.
\item \strong{"last"}: return rows that represent last attempt of each students.
\item \strong{"all"}: no filter applied, return all rows after applying \code{choose_encode}
}}

\item{split_cloze}{(Logical) If the \code{data} has at least one response column(s) that contained \href{https://docs.moodle.org/311/en/Embedded_Answers_(Cloze)_question_type}{embedded answers (Cloze)}, \code{split_cloze = TRUE} will split them into different parts (one column for each).
If the \code{data} has no Cloze response columns, \code{split_cloze = TRUE} will result in error message. If \code{FALSE} (default), no Cloze column will be split.}

\item{part_glue}{(Character) If \code{split_cloze = TRUE} and cloze response columns is presented, \code{part_glue} represent a character that \emph{glues} the names of "Response" column and each "Parts" of Cloze columns together.}

\item{sep_col}{(Character) If \code{data} is a named list of data.frame, \code{sep_col} indicate a character separation between names of list and "State" or "Response" columns.}
}
\value{
\strong{A data.frame}, its output content is determined by class of its first argument: \code{data}.
\itemize{
\item If the \code{data} is a data.frame; the output is an encoded, filtered, and cleaned data.frame of Moodle Responses report.
\item If the \code{data} is a named list of data.frame; the output is the same as previously described, but all Moodle Quiz reports are \href{https://dplyr.tidyverse.org/reference/mutate-joins.html}{full-joined} together by column "Name" and "ID".
So that, "Response" columns from each data.frame are sit together in a single data.frame. The resulting response columns are in this format (default): "<listNames>-Responses-<number>"
}
}
\description{
\code{\link[=combine_resp]{combine_resp()}} is a generic function that combine student responses from \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Responses report(s)} (not Grades report).
If the moodle responses report has response column(s) that contained \href{https://docs.moodle.org/311/en/Embedded_Answers_(Cloze)_question_type}{embedded answers (Cloze)}, you have an option to split them into different parts (one column for each).
Other functionality are similar to \code{\link[=check_sub]{check_sub()}} such as: encode & filter student's attempts by state (i.e., "Finished" or "In progress") and started time ("Started on" column),
cleans column names for easy manipulation, extracts student ID from "Email address", and unites "First name" and "Surname" column into "Name".
}
\examples{
NULL
}
