% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_answers.R
\name{count_answers}
\alias{count_answers}
\title{Count Student's Answers}
\usage{
count_answers(
  data,
  count_type = c("auto", "single", "multi", "cloze"),
  sort = TRUE,
  round_digits = 2
)
}
\arguments{
\item{data}{A data.frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Responses report(s)} (not Grades report)}

\item{count_type}{A character indicate type of counting methods. Must be one of:
\itemize{
\item \strong{"auto"} (default): This function detects the type of questions automatically and choose the best counting method it thinks for each type.
\item \strong{"single"}: It will simply count each answers by assuming that the question is a single answer type.
If the question is multiple answers type, counts of every combinations of the answers will be returned.
\item \strong{"multi"}: If the question is multiple answers type, this will give the count be each answers.
\item \strong{"cloze"}: If the question is embedded answers (Cloze), this will give the count by every "Parts" of the Cloze question pooled together.
}}

\item{sort}{(Logical) \code{TRUE}: sort the counted results}

\item{round_digits}{Integer to round the counted percentage or \code{NULL} to not round.}
}
\value{
A data.frame
}
\description{
\code{\link[=count_answers]{count_answers()}} counts student's answers from every "Response" columns from from a \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Responses report(s)}.
This is a wrapper around \code{\link[=count_answers_col]{count_answers_col()}} that counts from one column.
Answers from Moodle's quiz can be one of 3 types: single answered, multiple answered, or \href{https://docs.moodle.org/311/en/Embedded_Answers_(Cloze)_question_type}{embedded answers (Cloze)}.
Each of these answers types formatted differently in the "Response" column.
\code{\link[=count_answers]{count_answers()}} knows each type of answers and count the answers by the best counting method it thinks.
}
\examples{
NULL
}
