% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_answers.R
\name{count_answers_cloze_col}
\alias{count_answers_cloze_col}
\title{Count Answers from Cloze column}
\usage{
count_answers_cloze_col(data, col, count_cloze_parts = FALSE, sort = TRUE)
}
\arguments{
\item{data}{A data.frame of Moodle Responses report}

\item{col}{(quoted or unquoted) A column that contain Cloze answers}

\item{count_cloze_parts}{(Logical) If \code{TRUE}: count answers in each parts of Cloze column.}

\item{sort}{(Logical) sorted or not}
}
\value{
A data.frame
}
\description{
Count Answers from Cloze column
}
