% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_answers.R
\name{count_answers_col}
\alias{count_answers_col}
\title{Count Student's Answers from a Response Column}
\usage{
count_answers_col(
  data,
  col,
  count_type = c("auto", "single", "multi", "cloze"),
  count_cloze_parts = FALSE,
  sort = TRUE,
  round_digits = 2
)
}
\arguments{
\item{data}{A data.frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Responses report(s)} (not Grades report)}

\item{col}{Quoted or unquoted response column name that you want to count.}

\item{count_type}{A character indicate type of counting methods. Must be one of:
\itemize{
\item \strong{"auto"} (default): This function detects the type of question automatically and choose the best counting method it thinks.
\item \strong{"single"}: It will simply count each answers by assuming that the question is a single answer type.
If the question is multiple answers type, counts of every combinations of the answers will be returned.
\item \strong{"multi"}: If the question is multiple answers type, this will give the count be each answers.
\item \strong{"cloze"}: If the question is embedded answers (Cloze), this will give the count by every "Parts" of the Cloze question pooled together (default).
If \code{count_cloze_parts = TRUE}: it will count by each "Parts" of the Cloze question.
}}

\item{count_cloze_parts}{(Logical) Count each "Parts" of the Cloze question or not}

\item{sort}{(Logical) \code{TRUE}: sort the counted results}

\item{round_digits}{Integer to round the counted percentage or \code{NULL} to not round.}
}
\value{
A data.frame
}
\description{
Count student's answers from a given "Response" column from a \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Responses report(s)}. This function will count depending on
type of question. If the question is a single answer type, it will simply count the answers.
If the question is multiple answers type, it will count by each answers (default) or by combinations of answers.
If the question is \href{https://docs.moodle.org/311/en/Embedded_Answers_(Cloze)_question_type}{embedded answers (Cloze)},
it will count the answers by every "Parts" of the Cloze question pooled together (default), or by each "Parts" of the Cloze question.
}
\examples{
NULL
}
