% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_answers.R
\name{count_answers_multi_col}
\alias{count_answers_multi_col}
\title{Count Answers from Multiple Answered type question}
\usage{
count_answers_multi_col(data, col, sort = TRUE)
}
\arguments{
\item{data}{A data.frame of Moodle Responses report}

\item{col}{(quoted or unquoted) A column that contain Multiple Answered type question}

\item{sort}{(Logical) sorted or not}
}
\value{
A data.frame
}
\description{
Count Answers from Multiple Answered type question
}
