% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_moodle}
\alias{encode_moodle}
\title{Encode State column of Moodle Quiz report}
\usage{
encode_moodle(
  df_cleaned,
  state = c("Finished", "In progress"),
  encode = c(1, 0),
  choose_encode = c("max", "min", "all"),
  choose_time = c("first", "last", "all")
)
}
\arguments{
\item{df_cleaned}{A cleaned data.frame of Moodle Quiz report}

\item{state}{A character vector to match values in "State" column of Moodle Quiz report}

\item{encode}{An encoding numeric vector corresponding to \code{state}.
For example: by default, in the "State" column, "Finished" values will be encoded as \code{1}, and
"In progress" will be encoded as \code{0}.}

\item{choose_encode}{A character to filter student's attempt by the \code{encode}ing.
\itemize{
\item \strong{"max"} (default): return rows that have maximum encoding of each students.
\item \strong{"min"}: return rows that have mininum encoding of each students.
\item \strong{"all"}: no filter applied, return all rows.
}}

\item{choose_time}{A character to filter student's attempt by started time (determined by "Started on" column in Moodle Quiz report).
This filter applies \strong{after} \code{choose_encode} has been applied to the data.
\itemize{
\item \strong{"first"} (default): return rows that represent first attempt of each students.
\item \strong{"last"}: return rows that represent last attempt of each students.
\item \strong{"all"}: no filter applied, return all rows after applying \code{choose_encode}
}}
}
\value{
A data.frame that has been encoded and filtered
}
\description{
Encode State column of Moodle Quiz report
}
