% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encoder}
\alias{encoder}
\title{Encoding function}
\usage{
encoder(x, match, encode = match)
}
\arguments{
\item{x}{vector: input data to be matched.}

\item{match}{vector: matching value to \code{x}}

\item{encode}{vector: encoding vector same length as \code{match}}
}
\value{
encoded vector
}
\description{
Encoding function
}
\details{
Elements that not match will return \code{NA}.
}
