% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grades.R
\name{filter_grades_moodle}
\alias{filter_grades_moodle}
\title{Filter Grades}
\usage{
filter_grades_moodle(
  data_cleaned,
  choose_grade = c("max", "min", "mean", "all"),
  choose_time = c("first", "last", "all")
)
}
\arguments{
\item{data_cleaned}{A cleaned data.frame of Moodle Grades report}

\item{choose_grade}{(Character) To filter grade of each student}

\item{choose_time}{(Character) To filter attempt of each student}
}
\value{
A data.frame
}
\description{
Filter Grades
}
