% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_helper.R
\name{get_cloze_attr}
\alias{get_cloze_attr}
\title{Count Individual Parts of Cloze}
\usage{
get_cloze_attr(data)
}
\arguments{
\item{data}{A data.frame of Moodle Responses report}
}
\value{
A data.frame with 2 columns: \strong{\code{cloze_colnm}} for Cloze column names, and \strong{\code{parts}} for number of parts for each Cloze answers.
}
\description{
Count Individual Parts of Cloze
}
