% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_helper.R
\name{get_cloze_col_names}
\alias{get_cloze_col_names}
\title{Get Cloze column names}
\usage{
get_cloze_col_names(df)
}
\arguments{
\item{df}{A data.frame of Moodle Responses report}
}
\value{
Character vector indicate Cloze column names. If no cloze column, return \code{NULL}.
}
\description{
Get Cloze column names
}
