% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_helper.R
\name{get_cloze_parts}
\alias{get_cloze_parts}
\title{Get Cloze parts for 1 Cloze column}
\usage{
get_cloze_parts(data, cloze_col)
}
\arguments{
\item{data}{A data.frame of Moodle Responses report}

\item{cloze_col}{(character) Specify Cloze column to count parts.}
}
\value{
Numeric: indicate total parts of single Cloze answer.
}
\description{
Get Cloze parts for 1 Cloze column
}
