% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_attr.R
\name{get_max_grade}
\alias{get_max_grade}
\title{Get Maximum Grade of Quiz}
\usage{
get_max_grade(df_raw)
}
\arguments{
\item{df_raw}{A data.frame of Moodle Quiz report}
}
\value{
Numeric vector of length 1 indicate maximum quiz grade
}
\description{
Get a maxium grade of any Moodle quiz report.
}
