% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_attr.R
\name{get_max_resp}
\alias{get_max_resp}
\title{Get Maxiumn Responses of Moodle Responses report}
\usage{
get_max_resp(df_resp, count_cloze_parts = F)
}
\arguments{
\item{df_resp}{A data.frame of Moodle Responses report}

\item{count_cloze_parts}{(Logical) If \code{TRUE}, counts each parts of Cloze columns.
If \code{FALSE}, count each Cloze column as 1.}
}
\value{
An Integer
}
\description{
Get Maxiumn Responses of Moodle Responses report
}
