% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_helper.R
\name{get_noncloze_resp_colnm}
\alias{get_noncloze_resp_colnm}
\title{Get Non-Cloze Responses column name}
\usage{
get_noncloze_resp_colnm(df)
}
\arguments{
\item{df}{A data.frame of Moodle Responses report}
}
\value{
Character vector indicate Non-Cloze responses column names. If all responses are cloze column, return \code{NULL}.
}
\description{
Get Non-Cloze Responses column name
}
