% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_attr.R
\name{get_questions_no_max}
\alias{get_questions_no_max}
\title{Get Question Number and Maxiumn Score}
\usage{
get_questions_no_max(df_gr)
}
\arguments{
\item{df_gr}{A data.frame of Moodle Grades report}
}
\value{
A data.frame with 2 columns: \code{q_no} for questions number, and \code{q_max} for maximum score.
}
\description{
For Moodle Grades file, get a question number and corresponding question's maximum score.
}
