% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_attr.R
\name{get_responses_no}
\alias{get_responses_no}
\title{Get Responses Number}
\usage{
get_responses_no(df_resp)
}
\arguments{
\item{df_resp}{A data.frame of Moodle Response report}
}
\value{
Numeric vector
}
\description{
For Moodle Responses file, get a numeric vector for responses number.
}
