% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{grades_ls}
\alias{grades_ls}
\title{Moodle Grades Report Example}
\format{
A named list of 3 data.frame each has the following columns:
\itemize{
\item \strong{"Surname"}: surname of students (not real)
\item \strong{"First name"}: first name of students (not real)
\item \strong{"Institution"}: \code{NA}s
\item \strong{"Department"}: \code{NA}s
\item \strong{"Email address"}: Student's email address that has numeric student ID in it.
\item \strong{"State"}: State of the quiz
\item \strong{"Started on"}: Quiz start time
\item \strong{"Time taken"}: Elapsed time
\item \strong{"Completed"}: Quiz end time
\item \strong{"Grade/x"}: Student's score with "x" to indicate quiz maximum
\item \strong{"Q. n /x"}: Score of each questions; "n" indicate question number, "x" indicate maximum of each questions.
}
}
\usage{
grades_ls
}
\description{
Example for \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Grades report}.
Student's scores were retrieved form real world data, but student names, email address and time are randomly
simulated.
}
\keyword{datasets}
