% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_helper.R
\name{has_cloze_col}
\alias{has_cloze_col}
\title{Has any Cloze column?}
\usage{
has_cloze_col(df)
}
\arguments{
\item{df}{A data.frame of Moodle Responses report}
}
\value{
Logical: \code{TRUE} if \code{df} has at least one Cloze column.
}
\description{
Has any Cloze column?
}
