% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_helper.R
\name{is_cloze_col}
\alias{is_cloze_col}
\title{Is A Column contains Cloze Answers}
\usage{
is_cloze_col(data, col)
}
\arguments{
\item{data}{A data.frame of Moodle Responses report}

\item{col}{(tidy-select) column to test}
}
\value{
Logical: \code{TRUE} if \code{col} is a Cloze column.
}
\description{
Is A Column contains Cloze Answers
}
