% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quiz_type.R
\name{is_grades_report}
\alias{is_grades_report}
\title{Is it a Moodle Grades Report?}
\usage{
is_grades_report(data)
}
\arguments{
\item{data}{A data.frame to test.}
}
\value{
logical: \code{TRUE} if it is a Moodle Grades Report
}
\description{
Is it a Moodle Grades Report?
}
