% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_misc.R
\name{is_multi_resp_col}
\alias{is_multi_resp_col}
\title{Is a Column contained multiple answered MCQ}
\usage{
is_multi_resp_col(data, col)
}
\arguments{
\item{data}{A data.frame of Moodle Responses report}

\item{col}{(tidy-select) Column to test}
}
\value{
Named logical vector: \code{TRUE} if a column is a multiple answered MCQ
}
\description{
Is a Column contained multiple answered MCQ
}
