% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ls_df.R
\name{is_named_list_data.frame}
\alias{is_named_list_data.frame}
\title{Is an object is a Named List of Data Frame?}
\usage{
is_named_list_data.frame(x, quiet = FALSE)
}
\arguments{
\item{x}{An object to test}

\item{quiet}{If \code{quiet = TRUE} no message display}
}
\value{
logical
}
\description{
This function test an object if it is a Named list of data.frame
}
