% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_regex_in_cols}
\alias{is_regex_in_cols}
\title{Is Regular Expression presented in given column(s)}
\usage{
is_regex_in_cols(
  data,
  cols = tidyselect::everything(),
  regex,
  match_rows = c("any", "all"),
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data.frame}

\item{cols}{(tidy-select) columns to test}

\item{regex}{Character vector, specify regular expressions}

\item{match_rows}{(Character) one of "any": any rows that can match \code{regex}, "all": all rows must match \code{regex}}

\item{na.rm}{Logical, Passed to \code{match_rows}}
}
\value{
Named logical vector
}
\description{
Is Regular Expression presented in given column(s)
}
