% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_regex_in_names}
\alias{is_regex_in_names}
\title{Is Regular Expressions presented in object names}
\usage{
is_regex_in_names(x, regex, verbose = F)
}
\arguments{
\item{x}{An object to test}

\item{regex}{Character vector, specify regular expressions}

\item{verbose}{If \code{TRUE}, message you that in which components of the object names is/are not match by \code{regex}}
}
\value{
Logical, if \code{TRUE} all of the \code{regex} can be matched to at least one element of names of \code{x}.
}
\description{
Vectorized testing for regular expressions.
Are all of the regular expression can be matched to object names or not?
}
