% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quiz_type.R
\name{is_report}
\alias{is_report}
\title{Is it a Moodle Quiz Report?}
\usage{
is_report(data)
}
\arguments{
\item{data}{A data.frame to test.}
}
\value{
logical: \code{TRUE} if it is a Moodle Quiz Report
}
\description{
Is it a Moodle Quiz Report?
}
