% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quiz_type.R
\name{is_responses_report}
\alias{is_responses_report}
\title{Is it a Moodle Responses Report?}
\usage{
is_responses_report(data)
}
\arguments{
\item{data}{A data.frame to test.}
}
\value{
logical: \code{TRUE} if it is a Moodle Responses Report
}
\description{
Is it a Moodle Responses Report?
}
