% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_grade_col.R
\name{is_some_grade_numeric}
\alias{is_some_grade_numeric}
\title{Is some Grade Numeric ?}
\usage{
is_some_grade_numeric(data)
}
\arguments{
\item{data}{A data.frame to test.}
}
\value{
Logical: \code{TRUE} if numeric grade is presented.
}
\description{
Check in Grade column of the Moodle Quiz report. Are there any Numeric Grade?
}
