% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_grade_col.R
\name{is_some_grade_nyg}
\alias{is_some_grade_nyg}
\title{Is some Grade "Not yet graded"}
\usage{
is_some_grade_nyg(data)
}
\arguments{
\item{data}{A data.frame to test.}
}
\value{
Logical: \code{TRUE} if "Not yet graded" is presented.
}
\description{
Check in Grade column of the Moodle Quiz report. Are there any "Not yet graded"?
}
