% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleQuiz-package.R
\docType{package}
\name{moodleQuiz-package}
\alias{moodleQuiz}
\alias{moodleQuiz-package}
\title{moodleQuiz: A High-level R Package for Manipulation of Moodle Quiz Report}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide high-level ready-to-use functions for data wrangling (i.e., cleaning, filtering, and encoding) and combining Moodle Quiz Report (Grades Report and Responses Report).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lightbridge-KS/moodleQuiz}
  \item \url{https://lightbridge-ks.github.io/moodleQuiz/}
  \item Report bugs at \url{https://github.com/Lightbridge-KS/moodleQuiz/issues}
}

}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{kittipos.sir@mahidol.edu}

}
\keyword{internal}
