% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_msg}
\alias{print_msg}
\title{Print Message by a Separator}
\usage{
print_msg(x, ..., sep = "\\n", domain = NULL, appendLF = TRUE)
}
\arguments{
\item{x}{A character vector contained messages to print}

\item{...}{Pass to \code{...} of \code{message()}}

\item{sep}{Separator}

\item{domain}{Arg of \code{message()}}

\item{appendLF}{Arg of \code{message()}}
}
\value{
Messages
}
\description{
Print message from character vector by a given separator.
}
