% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_with_ls_df_names}
\alias{rename_with_ls_df_names}
\title{Prefix or Suffix column names in a list of Data frame with list names}
\usage{
rename_with_ls_df_names(
  ls_df,
  .cols,
  name_position = c("prefix", "suffix"),
  sep = "_"
)
}
\arguments{
\item{ls_df}{A list of data.frame}

\item{.cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to rename, this column must presented in all data.frame}

\item{name_position}{(Character) "prefix": prefix names to columns or "suffix": suffix names to columns}

\item{sep}{(Character) Indicate separation of names of list of data.frame and \code{.cols}}
}
\value{
A list of data.frame with renamed columns
}
\description{
Add names of list of data frame to prefix or suffix specified column names in data frames in a list of data frame.
}
