% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{responses_ls}
\alias{responses_ls}
\title{Moodle Responses Report Example}
\format{
A named list of 2 data.frame each has the following columns:
\itemize{
\item \strong{"Surname"}: surname of students (not real)
\item \strong{"First name"}: first name of students (not real)
\item \strong{"Institution"}: \code{NA}s
\item \strong{"Department"}: \code{NA}s
\item \strong{"Email address"}: Student's email address that has numeric student ID in it.
\item \strong{"State"}: State of the quiz
\item \strong{"Started on"}: Quiz start time
\item \strong{"Time taken"}: Elapsed time
\item \strong{"Completed"}: Quiz end time
\item \strong{"Grade/x"}: Student's score with "x" to indicate quiz maximum
\item \strong{"Response x"}: Essay responses of each student
}
}
\usage{
responses_ls
}
\description{
Example for \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Responses report}.
This data is entirely simulated. The texts in \verb{Response x} columns were randomly selected from "The Return of Tarzan" by Burroughs, Edgar Rice.
}
\keyword{datasets}
