% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_split.R
\name{split_cloze}
\alias{split_cloze}
\title{Detect and Split all Cloze columns in a Data Frame}
\usage{
split_cloze(data, part_glue = "_part_")
}
\arguments{
\item{data}{A data.frame of Moodle Responses report that has at least one embedded answer (Cloze) responses column.}

\item{part_glue}{(Character) This indicate characters that separate different "parts" of a Cloze column.}
}
\value{
A data.frame with cloze column splitted into multiple parts
}
\description{
This is a higher level function than \code{\link[=split_cloze_col]{split_cloze_col()}}, because it automatically finds all
embedded answers (Cloze) response columns and split all of them into multiple parts (one column for each part).
}
