% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze_split.R
\name{split_cloze_col}
\alias{split_cloze_col}
\title{Split Single Cloze Column}
\usage{
split_cloze_col(
  data,
  col,
  split_type = c("cols", "rows"),
  part_glue = "_part_"
)
}
\arguments{
\item{data}{A data.frame of Moodle Responses report that has at least one embedded answer (Cloze) responses column.}

\item{col}{Quoted or unquoted expression that indicate the Cloze response column name.}

\item{split_type}{(Character) Indicate output type: "cols" to split into multiple columns, "rows" to split into multiple rows.}

\item{part_glue}{(Character) If \code{split_type = "cols"}, this indicate characters that separate different "parts" of a Cloze column.}
}
\value{
A data.frame with cloze column splitted into multiple parts
}
\description{
An embedded answer (Cloze) response column usually has multiple parts separate by colon and semicolon
such as: "part 1: answer 1; part 2: answer 2;". This function split a single Cloze response column into multiple parts.
}
