% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze.R
\name{choices}
\alias{choices}
\title{Create a set of choices for single or multiple choice questions}
\usage{
choices(options, answer)
}
\arguments{
\item{options}{A character vector of selectable choices}

\item{answer}{A character vector of the correct answers}
}
\description{
Create a set of choices for single or multiple choice questions
}
