% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_quiz_type.R
\name{get_report_type}
\alias{get_report_type}
\title{Get Type of Moodle Quiz Report}
\usage{
get_report_type(data)
}
\arguments{
\item{data}{A data.frame to test.}
}
\value{
Character vector: \code{"Grades"} for Moodle Grades Report, \code{Responses}  for Moodle Responses Report,
or \code{NA} if it's not a Moodle Quiz report
}
\description{
Get Type of Moodle Quiz Report
}
