% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleQuiz-package.R
\docType{package}
\name{moodleQuiz-package}
\alias{moodleQuiz}
\alias{moodleQuiz-package}
\title{moodleQuiz: High-level functions for Wrangling, Combine, and Analyse Moodle Quiz Report}
\description{
This package provides high-level ready to use function that performs data wrangling, aggregration, and analysis on Moodle Quiz Report (i.e., Grades Report and Responses Report).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lightbridge-KS/moodleQuiz}
  \item Report bugs at \url{https://github.com/Lightbridge-KS/moodleQuiz/issues}
}

}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{kittipos.sir@mahidol.edu}

}
\keyword{internal}
